read/write ret code fixed. "samples/net" fixed too

+++ api/zrt.c
@@ -135,7 +135,7 @@ static int32_t zrt_read(uint32_t *args)
   if(length > 0) pos_ptr[file] += length;
 
   zrt_log("%lld bytes has been read", length);
-  return length;
+  return (int16_t)length;
 }
 
 /* write the file with the given handle number */
@@ -152,7 +152,7 @@ static int32_t zrt_write(uint32_t *args)
   if(length > 0) pos_ptr[file] += length;
 
   zrt_log("%lld bytes has been written", length);
-  return length;
+  return (int32_t)length;
 }
 
 /*
